refactor: split sqlite wrapper from walletkit-db
