+++ README.md
@@ -108,7 +108,8 @@ cargo fmt -- --check
 WalletKit is broken down into separate crates, offering the following functionality.
 
 - `walletkit-core` - Enables basic usage of a World ID to generate ZKPs using different credentials.
-- `walletkit-db` - Encrypted on-device storage primitives for WalletKit: `SQLCipher` (`sqlite3mc`) wrapper, vault opener, content-addressed blob storage, sealed key envelope, and cross-process lock. Used by `walletkit-core` for credential storage and consumable by sibling SDKs (e.g. OrbKit's planned `OrbPcpStore`).
+- `walletkit-sqlite` - Low-level safe Rust wrapper around encrypted SQLite (`sqlite3mc`), including connections, statements, transactions, and cipher operations.
+- `walletkit-db` - Higher-level encrypted storage abstractions for WalletKit: vault opener, content-addressed blob storage, sealed key envelope, and cross-process lock. Built on `walletkit-sqlite` and consumable by `walletkit-core` and sibling SDKs.
 
 ### World ID Secret
