+++ crates/walletkit-core/src/storage/cache/merkle.rs
@@ -1,7 +1,7 @@
 //! Merkle proof cache helpers.
 
 use crate::storage::{cache::schema::CACHE_KEY_PREFIX_MERKLE, error::StorageResult};
-use walletkit_db::Connection;
+use walletkit_sqlite::Connection;
 
 use super::util::{
     cache_entry_times, get_cache_entry, prune_expired_entries, upsert_cache_entry,
