+++ crates/walletkit-core/src/storage/cache/nullifiers.rs
@@ -7,7 +7,7 @@
 //!
 
 use crate::storage::error::StorageResult;
-use walletkit_db::Connection;
+use walletkit_sqlite::Connection;
 
 use super::util::{
     cache_entry_times, get_cache_entry, get_cache_entry_tx, insert_cache_entry_tx,
