+++ crates/walletkit-core/src/storage/cache/util.rs
@@ -6,7 +6,7 @@ use crate::storage::{
     cache::schema::{CACHE_KEY_PREFIX_REPLAY_NULLIFIER, CACHE_KEY_PREFIX_SESSION},
     error::{StorageError, StorageResult},
 };
-use walletkit_db::{params, Connection, DbError, Transaction};
+use walletkit_sqlite::{params, Connection, Error as DbError, Transaction};
 
 /// Maps a database error into a cache storage error.
 pub(super) fn map_db_err(err: &DbError) -> StorageError {
@@ -188,8 +188,8 @@ pub(super) fn get_cache_entry_tx(
         stmt.bind_values(params![key, now, insertion_before])
             .map_err(|err| map_db_err(&err))?;
         match stmt.step().map_err(|err| map_db_err(&err))? {
-            walletkit_db::StepResult::Row(row) => Ok(Some(row.column_blob(0))),
-            walletkit_db::StepResult::Done => Ok(None),
+            walletkit_sqlite::StepResult::Row(row) => Ok(Some(row.column_blob(0))),
+            walletkit_sqlite::StepResult::Done => Ok(None),
         }
     } else {
         let mut stmt = tx.prepare(
@@ -198,8 +198,8 @@ pub(super) fn get_cache_entry_tx(
         stmt.bind_values(params![key, now])
             .map_err(|err| map_db_err(&err))?;
         match stmt.step().map_err(|err| map_db_err(&err))? {
-            walletkit_db::StepResult::Row(row) => Ok(Some(row.column_blob(0))),
-            walletkit_db::StepResult::Done => Ok(None),
+            walletkit_sqlite::StepResult::Row(row) => Ok(Some(row.column_blob(0))),
+            walletkit_sqlite::StepResult::Done => Ok(None),
         }
     }
 }
