+++ crates/walletkit-db/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "walletkit-db"
-description = "Encrypted on-device storage primitives for WalletKit (SQLCipher wrapper, vault, content-addressed blobs, key envelope)."
+description = "Encrypted on-device storage abstractions for WalletKit (vault, content-addressed blobs, key envelope)."
 publish = true
 version.workspace = true
 edition.workspace = true
@@ -23,18 +23,13 @@ serde = { workspace = true, features = ["derive"] }
 sha2 = { workspace = true }
 thiserror = { workspace = true }
 zeroize = { workspace = true, features = ["derive"] }
+walletkit-sqlite = { workspace = true }
 
 [target.'cfg(target_os = "android")'.dependencies]
 sha2 = { workspace = true, features = ["force-soft"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { workspace = true, features = ["wasm_js"] }
-sqlite-wasm-rs = { workspace = true, features = ["sqlite3mc"] }
-
-[build-dependencies]
-cc = { workspace = true }
-sha2 = { workspace = true }
-zip = { workspace = true, features = ["deflate"] }
 
 [dev-dependencies]
 tempfile = { workspace = true }
