+++ crates/walletkit-db/src/lib.rs
@@ -3,10 +3,8 @@
 //! The crate provides building blocks shared by `walletkit-core::storage` and
 //! sibling SDKs (e.g. `OrbKit`'s `OrbPcpStore`):
 //!
-//! - [`Connection`], [`Transaction`], [`Statement`], [`cipher`] — encrypted
-//!   `SQLite` (`sqlite3mc`) wrapper with safe Rust types.
 //! - [`Vault`] — encrypted-database wrapper around a caller-supplied schema,
-//!   exposing the underlying [`Connection`].
+//!   exposing the underlying [`walletkit_sqlite::Connection`].
 //! - [`blobs`] — content-addressed blob storage (`ensure_schema`, `put`,
 //!   `get`), [`ContentId`], and [`compute_content_id`].
 //! - [`init_or_open_envelope_key`] — sealed intermediate key persisted via
@@ -25,7 +23,6 @@ pub mod blobs;
 mod envelope;
 mod error;
 mod lock;
-mod sqlite;
 mod traits;
 mod vault;
 
@@ -33,10 +30,6 @@ pub use blobs::{compute_content_id, ContentId};
 pub use envelope::init_or_open_envelope_key;
 pub use error::{StoreError, StoreResult};
 pub use lock::{Lock, LockGuard};
-pub use sqlite::{
-    cipher, Connection, DbResult, Error as DbError, Row, Statement, StepResult,
-    Transaction, Value,
-};
 pub use traits::{AtomicBlobStore, Keystore};
 pub use vault::Vault;
