+++ crates/walletkit-sqlite/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "walletkit-sqlite"
+description = "Safe Rust wrapper around sqlite3mc for WalletKit."
+publish = true
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+exclude.workspace = true
+readme = "README.md"
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+hex = { workspace = true }
+secrecy = { workspace = true }
+zeroize = { workspace = true }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+sqlite-wasm-rs = { workspace = true, features = ["sqlite3mc"] }
+
+[build-dependencies]
+cc = { workspace = true }
+sha2 = { workspace = true }
+zip = { workspace = true, features = ["deflate"] }
+
+[dev-dependencies]
+tempfile = { workspace = true }
+
+[lints]
+workspace = true
