+++ crates/walletkit-sqlite/README.md
@@ -1,0 +1,11 @@
+# walletkit-sqlite
+
+Low-level encrypted SQLite support for WalletKit.
+
+The crate provides safe Rust wrappers for connections, statements,
+transactions, values, and sqlite3mc cipher operations. On native targets it
+builds and statically links sqlite3mc; on WASM it delegates to
+`sqlite-wasm-rs`.
+
+It deliberately does not own schemas or storage policy. Those abstractions
+belong in `walletkit-db` or in its consumers.
