+++ crates/walletkit-sqlite/src/ffi.rs
